changing some colors + other misc changes
Changed the Hybrid.itermcolors so that terminal background is less
black.

Updated PS1 to use a unicode ❯ instead of the boring $.

Helpful little things like mapping H to ^ and L to $ for vim.

Removed a bunch of the spacemacs inspired vim bindings

Diff so fancy now shows + and - so that diffs are readable when flux is
on.

Added .ideavimrc. Source ~/.vimrc didn't work so I am just adding
specific config items.

Changed some tmux colors for selection & removed some unnecessary plugins
and bindings.

Added another script to the utils/color_text_test.bash